pyproject.toml
[project]
name = "fastapi-backend"
version = "0.0.1"
description = "A fastapi-backend Server for ICA"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.9",
    "pydantic>=2.11.6",
    "uvicorn>=0.34.0",
    "PyJWT>=2.10.0",
    "python-dotenv>=1.1.0",
    "openai>=1.97.0",
    "jinja2>=3.1.5",
    "requests>=2.32.4",
    "python-multipart>=0.0.20",
    "monday>=2.0.1"
]
